cloudtests: move exists and wait to util folder
