+++ test/cloudtest/test_antiaffinity.py
@@ -11,7 +11,7 @@ import pytest
 
 from materialize.cloudtest.app.materialize_application import MaterializeApplication
 from materialize.cloudtest.k8s import cluster_pod_name
-from materialize.cloudtest.wait import wait
+from materialize.cloudtest.util.wait import wait
 
 
 def zones_used(mz: MaterializeApplication) -> int:
