+++ test/cloudtest/test_crash.py
@@ -17,7 +17,7 @@ from pg8000.exceptions import InterfaceError
 
 from materialize.cloudtest.app.materialize_application import MaterializeApplication
 from materialize.cloudtest.k8s import cluster_pod_name
-from materialize.cloudtest.wait import wait
+from materialize.cloudtest.util.wait import wait
 
 
 def populate(mz: MaterializeApplication, seed: int) -> None:
