+++ test/cloudtest/test_managed_cluster.py
@@ -9,7 +9,7 @@
 
 from materialize.cloudtest.app.materialize_application import MaterializeApplication
 from materialize.cloudtest.k8s import cluster_pod_name
-from materialize.cloudtest.wait import wait
+from materialize.cloudtest.util.wait import wait
 
 
 def test_managed_cluster_sizing(mz: MaterializeApplication) -> None:
