+++ test/cloudtest/test_privatelink_connection.py
@@ -13,7 +13,7 @@ import pytest
 from pg8000.dbapi import ProgrammingError
 
 from materialize.cloudtest.app.materialize_application import MaterializeApplication
-from materialize.cloudtest.exists import exists, not_exists
+from materialize.cloudtest.util.exists import exists, not_exists
 
 
 def test_create_privatelink_connection(mz: MaterializeApplication) -> None:
