+++ test/cloudtest/test_secrets.py
@@ -15,7 +15,7 @@ from pg8000.exceptions import InterfaceError
 
 from materialize.cloudtest.app.materialize_application import MaterializeApplication
 from materialize.cloudtest.k8s import cluster_pod_name
-from materialize.cloudtest.wait import wait
+from materialize.cloudtest.util.wait import wait
 
 
 def test_secrets(mz: MaterializeApplication) -> None:
