+++ test/cloudtest/test_smoke.py
@@ -10,7 +10,7 @@
 from textwrap import dedent
 
 from materialize.cloudtest.app.materialize_application import MaterializeApplication
-from materialize.cloudtest.wait import wait
+from materialize.cloudtest.util.wait import wait
 
 
 def test_wait(mz: MaterializeApplication) -> None:
