+++ test/cloudtest/test_upgrade.py
@@ -18,7 +18,7 @@ from materialize.checks.cloudtest_actions import ReplaceEnvironmentdStatefulSet
 from materialize.checks.executors import CloudtestExecutor
 from materialize.checks.scenarios import Scenario
 from materialize.cloudtest.app.materialize_application import MaterializeApplication
-from materialize.cloudtest.wait import wait
+from materialize.cloudtest.util.wait import wait
 from materialize.util import MzVersion
 from materialize.version_list import VersionsFromDocs
